# backend/src/quotes/static/swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Quotes API
  description: API для обработки котировок
paths:
...
  /api/dwh:
    get:
      summary: Retrieve data from DWH.
      description: |
        Fetch records from the `dwh.data_mart` table with optional filters and pagination.
      tags:
        - DWH
      security:
        - BearerAuth: []
      parameters:
        - name: product_type
          in: query
          description: Filter by product type.
          schema:
            type: string
        - name: model_name
          in: query
          description: Filter by model name.
          schema:
            type: string
        - name: is_insurance_case
          in: query
          description: Filter by whether it is an insurance case.
          schema:
            type: boolean
        - name: feature_name
          in: query
          description: Filter by feature name.
          schema:
            type: string
        - name: start_date
          in: query
          description: Filter by start date (YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: Filter by end date (YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Number of records per page.
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful response with paginated data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    description: Current page number.
                  total_pages:
                    type: integer
                    description: Total number of pages.
                  total_records:
                    type: integer
                    description: Total number of records.
                  per_page:
                    type: integer
                    description: Number of records per page.
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        product_type:
                          type: string
                        run_id:
                          type: integer
                        quote_id:
                          type: integer
                        briefcase_date:
                          type: string
                          format: date
                        start_date:
                          type: string
                          format: date
                        end_date:
                          type: string
                          format: date
                        model_name:
                          type: string
                        feature_name:
                          type: string
                        feature_value:
                          type: string
                        predict:
                          type: number
                          format: float
                        is_insurance_case:
                          type: boolean
                        data_insurance_case:
                          type: string
        '400':
          description: Bad request due to invalid input.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      summary: Retrieve data from DWH with filters in request body.
      description: |
        Fetch records from the `dwh.data_mart` table by providing filters in the request body.
      tags:
        - DWH
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_type:
                  type: string
                model_name:
                  type: string
                is_insurance_case:
                  type: boolean
                feature_name:
                  type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '200':
          description: Successful response with paginated data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    description: Current page number.
                  total_pages:
                    type: integer
                    description: Total number of pages.
                  total_records:
                    type: integer
                    description: Total number of records.
                  per_page:
                    type: integer
                    description: Number of records per page.
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        product_type:
                          type: string
                        run_id:
                          type: integer
                        quote_id:
                          type: integer
                        briefcase_date:
                          type: string
                          format: date
                        start_date:
                          type: string
                          format: date
                        end_date:
                          type: string
                          format: date
                        model_name:
                          type: string
                        feature_name:
                          type: string
                        feature_value:
                          type: string
                        predict:
                          type: number
                          format: float
                        is_insurance_case:
                          type: boolean
                        data_insurance_case:
                          type: string
        '400':
          description: Bad request due to invalid input.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...